requirements.txt
# Core ML frameworks
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.20.0

# Fine-tuning frameworks
peft>=0.6.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
numpy>=1.24.0

# Optional logging (install with: pip install wandb)
# wandb>=0.15.0

# Development tools (optional)
# jupyter>=1.0.0
# matplotlib>=3.5.0
# seaborn>=0.11.0